Change default stack bounds behaviour
It now defaults to only setting bounds if needed (instead of all-or-nothing
as before) and creating new intrinsic calls unless there are 5 or more
users of the bounded capability in which case it will reuse a single
bounded intrinsic call.

This saves two instructions and one spilled stack capability in
webkit-inefficient.ll.