Ensure that underaligned memcpy() greater equal than CAP_SIZE uses a libcall
There was a bug where size == CAP_SIZE would inline the memcpy() if the
alignment is large enough to emit a sequence of library calls. Fixed by
using >= instead of > and forcing the use of a libcall by returning
MVT::isVoid instead of relying on SelectionDAG not expanding the memcpy
if we return MVT::i8 (this was true for CHERI256 but depending on the
threshold could expand the memcpy for CHERI128).
What we really want here is a clc/csc/memcpy instruction that can deal with
unaligned data and copies tags if it happens to be aligned.

This will probably make some benchmarks perform worse if they copy 16-byte
values that are known not to contain capabilities but there is no good
solution to this problem.