Stop using CSetOffset in cheri_init_globals.h
This is wrong! When using CHERI128 imprecise capabilities the csetoffset
on the bounded capabily might lose the last few bytes since the current
offset is not guaranteed to be zero after a csetbounds.

This was working fine with the 23-bit CHERI CC but with reduced precision
the offset would be non-zero more frequently and cause SPEC to fail because
some of the required global variables were not being initialized.