Revert "Avoid creating a capability store to an underaligned value in SROA"
Bisecting shows this as the underlying cause for WebKit crashing with a
tag violation. I'm not sure it is actually this commit but it seems to
be triggering something later on in the pipeline that goes away if I
disable the StackColoring pass. It is possible that it just reduces the
number of allocas so that stack-coloring no longer aliases two
allocations that are actually alive at the same time.

This reverts commit 382547d65bfe1b454f9c8b82fe28cb99e16e3460.