Initial support for lowering unaligned capability loads/stores
Convert it to a memcpy() to/from the stack instead of reporting a fatal
error. This way it will work at runtime if the values happen to be aligned.

Also emit a warning to indicate that something wrong is happening (should
probably be fixed by a cast to __uintcap_t*/__builtin_assume_aligned()).