[UPSTREAM][Verifier] Check that null call parameters are not marked as nonnull
I found that one of our out-of-tree optimization passes was incorrectly
replacing a nonnull argument with null but not removing the attribute.
I found this by chance while writing a new test. To avoid this in the
future I have now added a check in the verifier to catch such issues
earlier.