[subobject-bounds] Add a cheri_subobject_bounds_use_remaining_size attribute
If this attribute is applied to a field/type the bounds will be set to the
remaining bytes in the allocation (i.e. move cap.bottom to the start of the
field and keep top unchanged). The attribute also has an optional integer
argument which can be used as an upper bound for the size (e.g. 255 in
struct dirent).