[UPSTREAM][compiler-rt][fuzzer] Allow compiling for CHERI architecture
Almost all uses of uintptr_t only need a virtual address and not a pointer
sized integer (that for CHERI contains bounds and permissions, too).
For CHERI uintptr_t is 128 bits but virtual addresses are only 64 bits.
This change makes it a lot more efficient to use libfuzzer on CHERI.