[UPSTREAM][compiler-rt] Fix internal allocator alignment
The InternalAlloc() function would add a u64 after the real allocation and
then pass p + sizeof(u64) as the result. This causes the alignment value
to effectively be ignored and results in an alignment of only 8 bytes.
For CHERI this causes crashes almost immediately since loading/storing any
pointer on CHERI systems requires alignment of at least 16 bytes and will
trap otherwise.

This is probably not much of a problem on x86 but could cause some
performance degradation for larger types.

Currently, this causes another 64-byte field to be stored for every
allocation but this could be optimized in the future.

Furthermore, the default alignment for the internal allocator is now
max(8, sizeof(void*)) which will allow this allocator to be used for CHERI
and will keep the previous value of 8 on other architectures.