Drop CapWidth assertion from AsmPrinter
This embeds target-specific knowledge in the generic AsmPrinter, yet not
enough to be able to know what the right CapWidth is, only the set of
values. CHERI-64 will further add to this set, such that it will no
longer catch trying to emit address-sized quantities as capabilities on
64-bit architectures, so we should kill it and leave the target-specific
backend to perform any desired sanity checks, as already also happens.