[ELF] In the PLT ABI, we need ELF relocations even for static function pointers
Previously we were emitting __cap_relocs, i.e. a capability to the function.
This will not work if that function pointer is called from a different
context since it will not update $cgp.
This is slightly awkward since we can't add local symbols to the dynamic
symbol table in order to use them in a relocation.
Fix this by generating a new synthetic global (STV_HIDDEN) symbol and
using that instead.