Insert tail padding after global variables
This is conceptually a very simple change, but in order to get the CHERI
capability size to MCTargetStreamer, I had to modify a lot more code than
I would like. I think it should be possible to move things around a bit to
reduce the size of the diff but I'm not sure which MC class should hold the
capability size.