[subobject-bounds] Add a __builtin_no_change_bounds() expression
This can be used to opt-out of bounds for a given expression.
Note: it must wrap the expression that would have bounds added and not the
operation that causes the bounds to be added.

For example &__builtin_no_change_bounds(x->a) will retain the bounds of x
when taking the address of the member a, but the similar expression
__builtin_no_change_bounds(&x->a) will use tight bounds

TODO: diagnose those cases since they are almost certainly a typo.