Improve layout of C++ exception landing pad metadata
Emit a single zero byte for NULL values (ULEB128) and for non-NULL values
emit 0xc to indicate that the next capability sized and aligned value is
the real landing pad. The value 0xc could also be changed, I just used c
for capability.

This improves density of .gcc_except_table since we no longer waste up to
alignof(cap)-1+sizeof(cap)-1 bytes for NULL values and in most cases the 0xc
flag for capabilities will be within the range that needs to be padded for
alignment anyway.