[UPSTREAM?] MipsDelaySlotFiller: Don't move BUNDLE instructions into the delay slot
This breaks horribly since the BUNDLE will be removed and some of the
instructions end up in the next basic block. In order to ensure that
pc-relative address computations are correct (no instructions are inserted
between offset computation and getpcc), we bundle them together.
However, the delay-slot-filler broke this bundling and  caused interesting
failure modes when loading _CHERI_CAPABILITY_TABLE_ via $pcc.
In some functions this would move the computation of $cgp into a basic block
that is not always traversed before $cgp is dereferenced, resulting in
run-time crashes when using an uninitialized $cgp value.