Fix incorrect handling of __cheri_addr/__cheri_offset in IntExprEvaluator
We should be handling __cheri_addr in the same way as CK_PointerToIntegral.
For __cheri_offset we just give up and don't attempt to evaluate it.

I am not sure how this was not crashing in the past since the incorrect
return value (true instead of false) has been there from the beginning.

Fixes https://github.com/CTSRD-CHERI/llvm-project/issues/360