Infer the provenance source for addition and bitwise operations
If one of the sides of the expression is guaranteed to be a non-provenance-
carrying integer (e.g. a constant promoted to __uintcap_t), we derive the
result from the other operand.
If both operands could be a valid pointers cast to uintptr_t we now emit
a warning which can be silenced or promoted to an error with the
-Wcheri-provenance warning category.
If the provenance is ambiguous, we retain the current rule that it is
inherited from the LHS, but if this is rare, I may make the warning an
error by default.

For the implementation I add a new __attribute__((cheri_no_provenance)),
that can be used for fields, variables and typedefs to indicate the the
following variable cannot have CHERI tag bits.
In the future we might even add a UBSAN checker that ensures that all values
stored to such a variable are not tagged.
The compiler will infer this attribute for cast operations and propagate
it to the final result: e.g. `(uintptr_t)(void*)(uinptr_t)1` is known to
not carry provenance.

In this version we also infer the provenance source for multiplication
since it is a commutative operation. However, multiplying pointers doesn't
make much sense so I am not sure if this is needed.

Another questionable change in this patch is the addition of the inverse
attribute, cheri_carries_provenance, which is currently not used and I am
not sure if it will be needed (or is every useful) to resolve ambiguities.
I may be useful to add a compiler inserted check that each value stored
must be tagged, but even there I am unsure.

See https://github.com/CTSRD-CHERI/llvm-project/issues/340