Add a -cheri-comparison={address,exact} clang flag
This selects between ceq and cexeq code generation in the backend and will
allow the frontend to enable certain diagnostics only if exact equals is
enabled. One example where this is useful are the bitwise warnings
in address mode. Now that the comparision instructions no longer include
the tag bit, this warning should only be enabled if exact equals is being
used.