Skip the StackProtector pass when compiling purecap code
As the frontend should no longer be emitting functions with the ssp
attributes this should not make a difference, but maybe marginally improve
compilation time.

For testing purposes I added a flag --enable-purecap-stack-protector that
can be used to check that we can generate purecap code that uses stack
protectors (even though it's not useful in any way).