[ELF][Cheri] Remove assertion that non-TLS symbols aren't in TLS sections
The riscv-tests CRT code, which is copied in other riscv repositories,
uses _tdata_begin/_tdata_end/_tbss_end symbols defined in .tdata and
.tbss. These symbols are not thread-local and are intended to point at
the initialisation image, but nevertheless are in thread-local sections.
This is an odd case but BFD and LLD (without CHERI) do the right thing;
we should also permit this.