Disable GVN value coersion from CHERI capabilities
The check whether we can extract a given type at offset N from a capability
was wrong (and will need to be updated for RV32 and little-endian RV64).

As this is causing a serious miscompilation (#385), disable the optimization
until I have implemented the correct bit extraction patterns.
We can currently extract the address: this matches the second 64-bits
for MIPS and the first 64 for RISCV. We are also planning to add a
"extract high capability bits" instruction, so once we have that we can
extract all bits from capabilities and this code will become a lot simpler
and no longer require target-dependent checks.

Fixes https://github.com/CTSRD-CHERI/llvm-project/issues/385

This is a minor code-generation regression in the RV64 case, but I intend
to fix that in a follow-up commit.