[ELF] Only allow the emulation to enable CheriABI, not disable
On RISC-V (both bare-metal and FreeBSD), the emulation is always
provided, and without _fbsd/_cheri suffixes. The RISC-V approach is to
infer the ABI from the input files, and use one emulation per XLEN, so
let's just infer CheriABI if a non-suffixed emulation is provided. We
can probably also drop the _cheri suffix entirely for CHERI-MIPS now,
too.