Remove ancient legacy __cap_relocs code
Originally the compiler emitted the __cap_relocs section with relocations
for the individual fields. This kinda worked for some cases, but using
a single R_MIPS_CHERI_CAPABILITY (or the RISC-V equivalent) relocation is
vastly superior since it allows for sensible bounds.

This should have been deleted a long time ago.