Forbid cspecialw of cnull
This doesn't mean what it looks like, behaving like cspecialr cnull, scr
instead. That alias remains available for people who really want a
CSpecialRW that does nothing useful, as does the full cspecialrw form.
The latter is still dangerous, but we only define CSpecialRW in the spec
so it would be wrong to not permit giving it a source of cnull, and
would impose more complexity in splitting out cspecialr as its own
instruction rather than and alias (and then pose the question of whether
cspecialw should also be a separate instruction for uniformity). This
should help stop most instances of people shooting themselves in the
foot by not reading the spec properly.