Add a baseline test showing instructions being incorrectly hoisted
We should not be hoisting instructions that can potentially trap out of
basic blocks that might not be entered since this can result in unconditional
traps that would not normally have ocurred. Surprisingly this never
happened until we started enabling subobject bounds. With subobject bounds
some setbounds for subobjects are hoisted outside of a conditional basic
block and result in traps e.g. if the argument is NULL.

See https://github.com/CTSRD-CHERI/llvm-project/issues/359 and
https://github.com/CTSRD-CHERI/llvm-project/issues/348