Overload __builtin_cheri_cap_{to,from}_pointer
For from_pointer we accept all integer types except __uintcap_t as the
pointer argument in purecap and in hybrid we also allow all non-capability
pointers. If the argument is a pointer, the return type is the capability
equivalent of that pointer, if it's an integer it is void* __capability.

For to_pointer (only supported in hybrid) we accept all capability
pointer types and the return value is the non-capability equivalent of
that argument. If the argument is (u)intcap_t we return (u)intptr_t.