Make __intcap a proper keyword and primitive type
This mirrors how __int128 is implemented. The existing __intcap_t and
__uintcap_t names remain (and equivalents exist for __int128).

It's unclear what benefits this brings in practice, but aligning with
existing practices seems like a good idea.

There's a bunch of boring test churn here. It's debatable what should be
using __(u)intcap_t and what should be using (unsigned) __intcap, but a
mass replacement of the former with the latter in all the tests that are
broken due to the changes is simplest.