Update a test for CHERI-LLVM InstCombine difference
We don't move the GEP before the addrspacecast but it's not obvious why.
Our code is not incorrect, it just doesn't perform one of the transforms
that upstream does.
We should figure out why this is happening but this is low priority, so
for now just ensure that Jenkins is happy again.