[CHERI][SROA] Distinguish slices used to preserve capabilities
We were previously looking for splittable slices which read or write
tags to identify slices used to preseve capabilities. However the slices
can be made unsplittable, which mean this doesn't work.

Using the new flag allows us to correctly remove the 'dummy' slices
which don't have a corresponding read or write dummy slice.

Co-authored-by: Jessica Clarke <jrtc27@jrtc27.com>