Fix purecap unannotated cap-to-wider-than-address-int casts
For this case we were assuming that such casts would never be widening,
only the same or narrowing. Copy the CK_CHERICapabilityToAddress/Offset
code for this case instead so we handle it properly.

Only cap_to_oversized added in this commit was crashing, but the other
tests are added for completeness.