Add a -Wshorten-cap-to-int warning
This warning is intended to behave in the same way as -Wshorten-64-to-32
to catch cases where capability metadata is stripped at compile time
instead of at run time. Since this warning could be rather noisy it is
added the same group as -Wshorten-64-to-32 (-Wconversion).

This warning would have found various capability-stripping returns in
QtQml that I was only able to find at run time.