[SROA] Fix assertion on offset transfer with strict align slices
Due to the order in which we were inserting the strict alignment slices,
it could happen that the offset transfer logic in visitMemTransferInst()
sees a strict alignment slice instead of the actual slice covering the
full MemTransferInst. This can then result in various assertion failures
later on due to unexpected slice ranges. To fix this, we insert the
actual use first and then add the strict align slices afterwards.

This fixes at least two distinct assertion failures: I found the first one
while compiling Okular's synctex parser, the other one was reported as
a crash while compiling busybox awk.