Stop using __attribute__((cheri_capability)) in tests
This is an internal implementation detail. The only place this persists
in tests is checking what __capability is defined to, but for everything
else we should use __capability itself.

Note that memcpy-unaligned-nowarn-big.c has always been slightly wrong
since it was first committed, with the ((cheri_capability)) and space
missing, instead resulting in the first field being a plain pointer
called __attribute__lcsf_caller_pcc. This didn't actually matter though
because the second field padded and aligned the struct to the same as if
the first field were correctly defined, and nothing uses the fields,
they're just there to create a big enough and aligned enough struct.