Support __capability on array function parameters
For array parameters, type qualifiers go inside the square brackets and
qualify the pointer which the array decays to when used in semantic
analysis (note that the type is always the decayed form, unlike other
array uses). Thus, __capability should also be supported and mean that
the passed array is a capability pointer. This could not be done before
__capability was a keyword, as the parser rejects attributes inside the
square brackets.