Extend cheri-pointer-comparison-unsigned.ll to all comparisons and null checks
The existing tests are reordered so they now match the order of ICMP_FOO
in the enum.

Whilst here, remove the unnecessary uses of noreturn; the calls are
marked with tail and are in tail positions, so will be tail-called
already if supported by the backend (and if, like in the case of MIPS,
tail calls aren't supported, then noreturn won't change that).