[ConstantFold][GlobalOpt] Don't optimize non-zero i128->cap bitcasts
Trying to convert a load/store of an i128<->capbility bitcast was
previously trying to create a inttoptr/ptrtoint between the two. However,
we can't do that transformation without hardcoding detailed knowledge of
the capability format inside the ConstantFolding IR passs. As this is a
target-independent pass we should restrict it to zero constants in the
same way as it is currently done for non-integral pointers.