libunwind: Remove bogus downstream assert
This was added in 2017 in 4ae1b6a2998f without justification. There is
already an assertion upstream that UnwindCursor does not exceed
unw_cursor_t's size and, more recently, alignment, and the stronger
assertion that they're always equal does not hold on i386, where
unw_cursor_t is bigger than UnwindCursor. Not to mention that the assert
should have been written as an inline static_assert like the others, not
with this weird helper function.