Make sure to not break lifetime markers in certain edge cases
Currently, code such as:

  %0 = alloca
  %1 = bitcast ... %0 to i8 addrspace(200)*
  %2 = @llvm.lifetime.start.p200i8(..., %1)
  ...
  unsafe use of %1

causes problems for CheriBoundAllocas. Since the bitcast has an unsafe
use, we end up bounding the bitcast input (trying to maintain the
illusion that the alloca itself does the bounding). However, since in
this specific case the unsafe array uses the alloca as an i8 pointer,
just as the lifetime start/end arguments, they reuse the same bitcast,
and so we end up bounding the input to the lifetime markers.

We could try and copy the subtrees that have unsafe uses whilst leaving
the lifetime markers untouched, but that is complex and fragile.
Instead, for static allocas, we already have a special intrinsic we can
key off in ValueTracking so that the lifetime markers can peek through
the bounding, so do so, and add a separate one for dynamic allocas
(since some places assume that cheri_bounded_stack_cap will take a
FrameIndex operand).